+++ lsp/src/lsp.ml
@@ -10,3 +10,4 @@ module Text_document = Text_document
 module Types = Types
 module Uri = Uri0
 module Io = Io
+module Cli = Cli