+++ ocaml-lsp-server/bin/main.ml
@@ -1,16 +1,35 @@
+open Stdune
+module Cli = Lsp.Cli
+
 let () =
   Printexc.record_backtrace true;
   let version = ref false in
   let read_dot_merlin = ref false in
-  Arg.parse
+  let arg = Lsp.Cli.Arg.create () in
+  let spec =
     [ ("--version", Arg.Set version, "print version")
     ; ( "--fallback-read-dot-merlin"
       , Arg.Set read_dot_merlin
       , "read Merlin config from .merlin files. The `dot-merlin-reader` \
          package must be installed" )
     ]
-    (fun _ -> raise (Arg.Bad "anonymous arguments are not accepted"))
-    "ocamllsp";
+    @ Cli.Arg.spec arg
+  in
+  let usage =
+    "ocamllsp [ --stdio | --socket SOCKET --port PORT | --pipe PIPE ]"
+  in
+  Arg.parse
+    spec
+    (fun _ -> raise @@ Arg.Bad "anonymous arguments aren't allowed")
+    usage;
+  let channel =
+    match Cli.Arg.read arg with
+    | Ok c -> c
+    | Error s ->
+      Format.eprintf "%s@.%!" s;
+      Arg.usage spec usage;
+      exit 1
+  in
   let version = !version in
   if version then
     let version = Ocaml_lsp_server.Version.get () in
@@ -19,7 +38,7 @@ let () =
     let module Exn_with_backtrace = Stdune.Exn_with_backtrace in
     match
       Exn_with_backtrace.try_with
-        (Ocaml_lsp_server.run ~read_dot_merlin:!read_dot_merlin)
+        (Ocaml_lsp_server.run channel ~read_dot_merlin:!read_dot_merlin)
     with
     | Ok () -> ()
     | Error exn ->