+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Features
 
+- Support connecting over pipes and socket. Pipes on Windows aren't yet
+  supported (#946)
+
 - Semantic highlighting support is enabled by default (#933)
 
 - Re-enable `ocamlformat-rpc` for formatting code snippets (but not files) (#920, #939)